online_inference/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
fastapi==0.65.1
h11==0.12.0
idna==2.10
iniconfig==1.1.1
joblib==1.0.1
marshmallow==3.12.1
marshmallow-dataclass==8.4.1
mypy-extensions==0.4.3
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
src==0.1.0
starlette==0.14.2
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
uvicorn==0.13.4
websocket-client==0.59.0
